Button fixes (#13)
* add textDecoration to tailwind config

* BREAKING CHANGE: button flexbox layout

Previously, the .btn class was `display: flex`, which works just fine
when it's locked in a table. However, most buttons don't expand to fill
their horizontal container, which calls for `display: inline-flex`. This
also means that the .btn-inline class is no longer necessary, since its
layout is now the default.

* fix: add no-underline to buttons

* fix: add box-sizing: border-box to * and .btn

* add more explicit button examples to fractal templates