Add npm dist-tags for each published branch
This PR adds another step after our npm publishing in GitHub Actions that [adds an npm dist-tag](https://docs.npmjs.com/adding-dist-tags-to-packages#publishing-a-package-with-a-dist-tag) to the released version in the form `branch/{branch}`. The idea here is that we can use the branch tags in formio-sfds and on sfgov to keep the work in sync.

There's an additional job that should clean up the dist-tags for deleted branches. We should be able to merge this to `main` and test the workflow from start to finish without disrupting other work.